revit-store/backend/requirements.txt
# FastAPI та веб-сервер
fastapi==0.109.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6

# База даних
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Автентифікація та безпека
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic[email]==2.5.3

# AWS S3
boto3==1.33.13
aioboto3==12.2.0

# HTTP клієнт
httpx==0.26.0
aiohttp==3.9.1

# Telegram
aiogram==3.3.0

# Утиліти
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.2.0

# Тестування (опціонально)
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS
fastapi-cors==0.0.6
